move service deletion code to another file
